[REFACTOR] Change Amount Type

+++ src/main/java/id/ac/ui/cs/advprog/bepayment/model/Wallet.java
@@ -20,9 +20,9 @@ public class Wallet {
     @Id
     private String id;
     private String userId;
-    private Integer amount;
+    private double amount;
 
-    public Wallet(String id, String userId, Integer amount) {
+    public Wallet(String id, String userId, double amount) {
         this.id = id;
         this.userId = userId;
         this.amount = amount;
